clean home page

+++ include/functions.inc.php
@@ -37,32 +37,42 @@ function errorElement($msg) {
 }
 
 function printCustomers() {
-    global $handle;
-    while($cust = $handle->fetch(PDO::FETCH_ASSOC)) {
-        $created_date = DateTime::createFromFormat('Y-m-d', $cust['created_at']);
-        $fmtcreatedDate = $created_date->format('d-m-Y');
-        $edited_date = DateTime::createFromFormat('Y-m-d', $cust['edited_at']);
-        $fmteditedDate = $edited_date->format('d-m-Y');
-        echo' 
-            <tr class="customer-table-row">
-                <td class="pad-large mobile-stay">'. $cust["cust_name"] . '</td>
-                <td>' . $cust['cust_zip'] . ' '. $cust['cust_city'] . ', ' . $cust['cust_street'] . '</td>
-                <td>' . $cust['cust_phone'] . '</td>
-                <td>' . $fmtcreatedDate . '</td>
-                <td>' . $fmteditedDate . '</td>
-                <td id="edit-btn" class="text-center mobile-stay">
-                    <a href="../auth/home.php?id='. $cust["cust_id"] .'">
-                        <i class="fa-solid fa-pencil hover-scale"></i>
-                    </a>
-                </td>
-                <td class="text-center mobile-stay">
-                    <a href="../auth/deleteCustomer.php?id='. $cust["cust_id"] .'">
-                        <i class="fa-solid fa-trash-can hover-scale"></i>
-                    </a>
-                </td>
-            </tr>
-        ';
+    global $pdo;
+    $sql = "SELECT  cust_name, cust_phone, cust_zip, cust_city, cust_street, created_at, edited_at, cust_id FROM customers WHERE created_from = :createdId";
+    $handle = $pdo->prepare($sql);
+    $params = ['createdId'=>$_SESSION['users_id']];
+    $handle->execute($params);
+    if($handle->rowCount() > 0){
+        $isEmptyCustomers = false;
+        while($cust = $handle->fetch(PDO::FETCH_ASSOC)) {
+            $created_date = DateTime::createFromFormat('Y-m-d', $cust['created_at']);
+            $fmtcreatedDate = $created_date->format('d-m-Y');
+            $edited_date = DateTime::createFromFormat('Y-m-d', $cust['edited_at']);
+            $fmteditedDate = $edited_date->format('d-m-Y');
+            echo' 
+                <tr class="customer-table-row">
+                    <td class="pad-large mobile-stay">'. $cust["cust_name"] . '</td>
+                    <td>' . $cust['cust_zip'] . ' '. $cust['cust_city'] . ', ' . $cust['cust_street'] . '</td>
+                    <td>' . $cust['cust_phone'] . '</td>
+                    <td>' . $fmtcreatedDate . '</td>
+                    <td>' . $fmteditedDate . '</td>
+                    <td id="edit-btn" class="text-center mobile-stay">
+                        <a href="../auth/home.php?id='. $cust["cust_id"] .'">
+                            <i class="fa-solid fa-pencil hover-scale"></i>
+                        </a>
+                    </td>
+                    <td class="text-center mobile-stay">
+                        <a href="../auth/deleteCustomer.php?id='. $cust["cust_id"] .'">
+                            <i class="fa-solid fa-trash-can hover-scale"></i>
+                        </a>
+                    </td>
+                </tr>
+            ';
+        }
+    } else {
+        $isEmptyCustomers = true;
     }
+
 }
 
 function getTotalEntries() :int {
